requirements.txt
0==0.0.0
alabaster==0.7.12
allennlp==0.8.0
attrs==21.4.0
aws-xray-sdk==0.95
awscli==1.22.41
Babel==2.9.1
backports.weakref==1.0rc1
bleach==1.5.0
blis==0.7.5
boto==2.49.0
boto3==1.20.41
botocore==1.23.41
Bottleneck==1.2.1
cached-property==1.5.2
catalogue==1.0.0
certifi==2020.6.20
cffi==1.11.5
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
conllu==0.11
cookies==2.2.1
cryptography==3.2.1
cycler==0.11.0
cymem==2.0.6
cytoolz==0.9.0.1
dataclasses==0.8
dill==0.2.9
docker==5.0.3
docutils==0.15.2
ecdsa==0.17.0
editdistance==0.6.0
en-core-web-sm @ file:///data1/tuyunbin/PR/data/en_core_web_sm-2.3.0.tar.gz
flaky==3.7.0
Flask==1.0.2
Flask-Cors==3.0.7
ftfy==6.0.3
future==0.18.2
gevent==1.3.6
greenlet==1.1.2
h5py==2.10.0
html5lib==0.9999999
idna==3.3
imagesize==1.3.0
importlib-metadata==4.8.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsondiff==1.1.1
jsonnet==0.10.0
jsonpickle==2.1.0
kiwisolver==1.3.1
Markdown==2.2.0
MarkupSafe==2.0.1
matplotlib==2.2.3
mock==4.0.3
moto==1.3.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.6
nltk==3.6.1
numpy==1.19.5
numpydoc==0.8.0
overrides==3.1.0
packaging==21.3
parsimonious==0.8.0
Pillow==8.4.0
plac==0.9.6
pluggy==1.0.0
preshed==3.0.6
protobuf==3.19.3
py==1.11.0
pyaml==21.10.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.13.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-jose==2.0.2
pytorch-pretrained-bert==0.3.0
pytz==2017.3
PyYAML==5.4.1
regex==2018.1.10
requests==2.27.1
responses==0.17.0
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==0.19.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
snowballstemmer==2.2.0
spacy==2.3.7
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.2.4
srsly==1.0.5
tensorboardX==1.2
tensorflow==1.2.0
thinc==7.4.5
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
torch==1.3.0
torchvision==0.2.0
tqdm==4.62.3
typing-utils==0.1.0
typing_extensions==4.0.1
ujson==4.3.0
Unidecode==1.3.2
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
websocket-client==1.2.3
Werkzeug==2.0.2
wrapt==1.10.11
xmltodict==0.12.0
zipp==3.6.0
